requirements.txt
apiai==1.2.3
appdirs==1.4.3
botimize==1.1
certifi==2017.4.17
future==0.16.0
gunicorn==19.7.1
numpy==1.12.1
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
urllib3==1.20
